+++ lib/string_ext.dart
@@ -1,1 +1,3 @@
 library string_ext;
+
+export 'src/string_extensions.dart';